Switch to using the refactored version of defense
